+++ ChangeLog.md
@@ -22,6 +22,8 @@ Initial release.
   - Plugin system for adding custom format converters.
   - Field validation and error handling to ensure data integrity during
     conversion.
+  - Logging system with colorized output and possibility to set log verbosity
+    via CLI.
 - Ability to specify default currency when it's not present in the input CSV.
 - Generic format plugin with flexible column arrangement and activity names.
   Mainly as an example, because Wealthfolio's import system is very flexible and